updated the products grid home page
