+++ src/index.css
@@ -1,7 +1,22 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300;500;900&display=swap");
 *,
 *::before,
 *::after {
 	padding: 0;
 	margin: 0;
 	box-sizing: border-box;
+}
+
+body {
+	font-family: "Roboto", sans-serif;
+}
+
+a {
+	text-decoration: none;
+	color: #636363;
+}
+
+a:visited,
+a:focus {
+	color: #636363;
 }
