+++ src/routes/navigation/navigation.css
@@ -23,4 +23,5 @@
 .navigation .nav-link {
       padding: 10px 15px;
       cursor: pointer;
+      font-family: 'Roboto', sans-serif;
 }
