+++ src/styles/Category.css
@@ -1,68 +1,80 @@
-.category-container {
-  min-width: 30%;
-  height: 240px;
-  flex: 1 1 auto;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid black;
-  margin: 0 7.5px 15px;
-  overflow: hidden;
+.categories-container {
+	width: 100%;
+	display: flex;
+	flex-wrap: wrap;
+	justify-content: space-between;
 }
 
-.category-container .background-image {
-  transform: scale(1.1);
-  transition: transform 6s cubic-bezier(0.25, 0.45, 0.45, 0.95);
+.category-container {
+	min-width: 30%;
+	height: 400px;
+	flex: 1 1 auto;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	border: none;
+	margin: 0 7.5px 15px;
+	overflow: hidden;
+}
+
+
+.background-image:hover {
+	transform: scale(1.1);
+	transition: transform 5s cubic-bezier(0.25, 0.45, 0.45, 0.95);
 }
 
 .category-container:hover {
-  cursor: pointer;
+	cursor: pointer;
 }
 
-.category-body-container {
-  opacity: 0.9;
+.category-body {
+	opacity: 0.9;
 }
 
-.category-body-container.large {
-    height: 380px;
+.category-body.large {
+	height: 380px;
 }
 
-.category-body-container:first-child {
-    margin-right: 7.5px;
+.category-body:first-child {
+	margin-right: 7.5px;
 }
 
-.category-body-container:last-child {
-    margin-left: 7.5px;
+.category-body:last-child {
+	margin-left: 7.5px;
 }
 
-.category-body-container.background-image {
-  width: 100%;
-  height: 100%;
-  background-size: cover;
-  background-position: center;
+.background-image {
+	width: 100%;
+	height: 100%;
+	background-size: cover;
+	background-position: center;
+	object-fit: cover;
 }
 
-.category-body-container {
-  height: 90px;
-  padding: 0 25px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid black;
-  background-color: white;
-  opacity: 0.7;
-  position: absolute;
+.category-body {
+	height: 100px;
+	padding: 0 25px;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	border: 1px solid black;
+	background-color: white;
+	opacity: 0.7;
+	position: absolute;
 }
 
-.category-body-container h2 {
-  font-weight: bold;
-  margin: 0 6px 0;
-  font-size: 22px;
-  color: #4a4a4a;
+.category-body h2 {
+	font-weight: 600;
+  letter-spacing: 3px;
+  margin-bottom: 10px;
+	margin: 0 6px 0;
+	font-size: 22px;
+	color: #4a4a4a;
+  text-transform: uppercase;
 }
 
-.category-body-container p {
-      font-weight: lighter;
-      font-size: 16px;
+.category-body p {
+	font-weight: lighter;
+	font-size: 16px;
 }
